style - 파일 분리 및 가독성 개선
