+++ data/src/main/java/com/yapp/growth/data/internal/response/UserResponseImpl.kt
@@ -1,0 +1,8 @@
+package com.yapp.growth.data.internal.response
+
+import com.yapp.growth.data.response.UserResponse
+
+data class UserResponseImpl (
+    override val id: Long,
+    override val userName: String
+): UserResponse
