+++ data/src/main/java/com/yapp/growth/data/response/CategoryResponse.kt
@@ -1,0 +1,6 @@
+package com.yapp.growth.data.response
+
+interface CategoryResponse {
+    val id: Long
+    val keyword: String
+}
