+++ data/src/main/java/com/yapp/growth/data/response/UserResponse.kt
@@ -1,0 +1,6 @@
+package com.yapp.growth.data.response
+
+interface UserResponse {
+    val id: Long
+    val userName: String
+}
